+++ src/Group.h
@@ -41,6 +41,8 @@ protected:
     unsigned int maxPayload;
     Hub *hub;
     int extensionOptions;
+    int deflateWindowBits = 15;
+    int deflateMemLevel = 8;
     cS::Timer *timer = nullptr, *httpTimer = nullptr;
     const char *userPingMessage;
     size_t userPingMessageLength;
@@ -61,7 +63,7 @@ protected:
     void addHttpSocket(HttpSocket<isServer> *httpSocket);
     void removeHttpSocket(HttpSocket<isServer> *httpSocket);
 
-    Group(int extensionOptions, unsigned int maxPayload, Hub *hub, cS::NodeData *nodeData);
+    Group(int extensionOptions, unsigned int maxPayload, Hub *hub, cS::NodeData *nodeData, int deflateWindowBits = 15, int deflateMemLevel = 8);
     void stopListening();
 
 public: